main.py : Prediction with LBPH

+++ main.py
@@ -2,6 +2,7 @@
 from train import *
 import os 
 
+print cv2.__version__
 path = os.path.dirname(__file__)
 if len(path) != 0:
     os.chdir(os.path.dirname(path))
@@ -12,36 +13,37 @@ try:
 except:
     print 'err'
     pass
+capture = raw_input('Capture image for training[y/n]: ')
+if str(capture).lower() == 'y':
+    cap = cv2.VideoCapture(0)
+    count=0
+    while True:
 
-cap = cv2.VideoCapture(0)
-count=0
-while True:
+        _,frame = cap.read()
+        detectedFaces = extractFace(frame)
 
-    _,frame = cap.read()
-    detectedFaces = extractFace(frame)
+        if detectedFaces is None:
+            cv2.putText(frame,'Face Not Detected',(10,20),cv2.FONT_HERSHEY_COMPLEX_SMALL,1,(0,0,255),1)
+            cv2.imshow('Face Selector',frame)
+        else:
+            count+=1
+            face = cv2.resize(detectedFaces,(200,200))
+            face = cv2.cvtColor(face,cv2.COLOR_BGR2GRAY)
 
-    if detectedFaces is None:
-        cv2.putText(frame,'Face Not Detected',(10,20),cv2.FONT_HERSHEY_COMPLEX_SMALL,1,(0,0,255),1)
-        cv2.imshow('Face Selector',frame)
-    else:
-        count+=1
-        face = cv2.resize(detectedFaces,(200,200))
-        face = cv2.cvtColor(face,cv2.COLOR_BGR2GRAY)
+            file_name_path = './faces/users/'+str(count)+'.jpg'
+            cv2.imwrite(file_name_path,face)
 
-        file_name_path = './faces/users/'+str(count)+'.jpg'
-        cv2.imwrite(file_name_path,face)
+            cv2.putText(detectedFaces,str(count),(10,20),cv2.FONT_HERSHEY_COMPLEX_SMALL,1,(0,0,255),1)
+            cv2.imshow('Face Selector',detectedFaces)
 
-        cv2.putText(detectedFaces,str(count),(10,20),cv2.FONT_HERSHEY_COMPLEX_SMALL,1,(0,0,255),1)
-        cv2.imshow('Face Selector',detectedFaces)
-
-    k =cv2.waitKey(1) 
+        k =cv2.waitKey(1) 
 
 
-    if k==27 or count == 100 :
-        break
+        if k==27 or count == 300 :
+            break
 
-cv2.destroyAllWindows()
-cap.release()
+    cv2.destroyAllWindows()
+    cap.release()
 
 ##########################
 ##### TRAINING MODEL #####
@@ -49,17 +51,58 @@ cap.release()
 
 data_path = './faces/users/'
 
-files = [f for f in os.listdir(data_path) if isfile(join(data_path,f))]
+files = [f for f in os.listdir(data_path) if os.path.isfile(os.path.join(data_path,f))]
 
 trainingData,Labels = [],[]
 
 
-for i,files in enumerate(files):
+for i,f in enumerate(files):
     image_path = data_path + files[i]
     images = cv2.imread(image_path,0)
-    trainingData.append(np.asarray(images,dtype=uint8))
+    #print type(images)
+    trainingData.append(np.asarray(images,dtype=np.uint8))
     Labels.append(i)
 
 Labels =np.asarray(Labels,dtype=np.int32)
 
-model = cv2.face.createLBPHFaceRecognizer()
+model = cv2.face.LBPHFaceRecognizer_create()
+
+model.train(np.asarray(trainingData),np.asarray(Labels))
+
+
+############################
+###Run Facial Recognition###
+############################
+cap = cv2.VideoCapture(0)
+
+while True:
+    
+    ret,frame = cap.read()
+
+    try:
+        face = extractFace(frame)
+        face = cv2.resize(face,(200,200))
+        face = cv2.cvtColor(face,cv2.COLOR_BGR2GRAY)
+        results = model.predict(face)
+
+        if results[1] < 500:
+            confidence = int(100*(1-results[1]/300))
+            dis_str = str(confidence)+' confident'
+        
+        if confidence > 87:
+            cv2.putText(frame,dis_str,(100,50),cv2.FONT_HERSHEY_COMPLEX,1,(0,255,0),1)
+            cv2.imshow('Detection',frame)
+        else:
+            cv2.putText(frame,'Not Detected',(100,50),cv2.FONT_HERSHEY_COMPLEX,1,(0,255,0),1)
+            cv2.imshow('Detection',frame) 
+    except:
+        cv2.putText(frame,'No Face',(100,50),cv2.FONT_HERSHEY_COMPLEX,1,(0,255,0),1)
+        cv2.imshow('Detection',frame)
+
+    
+    k = cv2.waitKey(1)
+    if k==27:
+        break
+
+cap.release()
+cap.destroyAllWindows()
